requirements.txt
flask==2.3.3
requests==2.31.0
psutil==5.9.6
colorama==0.4.6
pyyaml==6.0.1
